Update Authorization and Timestamp usage
Followerwonk's Social Authority API was developed when it was a Moz
product. It used an Authorization method called MozSigned. Since
Followerwonk is no longer owned by, or associated with Moz, we've
renamed MozSigned to WonkSigned to avoid the confusion. We still accept
MozSigned, though, so existing code does not break.

The Timestamp parameter was originally called Expires, which allowed the
client to set the expiration policy. Long ago, while still a Moz
product, that strategy changed. Expires was renamed Timestamp, meant to
be the current time, allowing server-side request signature lifetime
policy management.

Just set Timestamp to the current time and if it is reasonably close to
current time when received by the server, it will be honored. A shorter
valid time window may be employed in the future.

Examples previously set a time in the future, compatible with the old
Expires policy. This update changes the examples to use current time
for all Timestamps.